Office/requirements.txt
inputimeout==1.0.4
natsort==8.4.0
openpyxl==3.1.2
pdf2image==1.16.3
Pillow==11.3.0
PyPDF2==3.0.1
python_docx==1.1.2
pypdf
